requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.3
curl_cffi==0.12.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
holidays==0.77
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
multitasking==0.0.12
namex==0.1.0
narwhals==2.0.1
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prophet==1.1.7
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stanio==0.5.1
streamlit==1.47.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yfinance==0.2.65
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.3
curl_cffi==0.12.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
holidays==0.77
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
multitasking==0.0.12
namex==0.1.0
narwhals==2.0.1
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.1.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prophet==1.1.7
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stanio==0.5.1
streamlit==1.47.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yfinance==0.2.65
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.3
curl_cffi==0.12.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
holidays==0.77
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
multitasking==0.0.12
namex==0.1.0
narwhals==2.0.1
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.1.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prophet==1.1.7
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stanio==0.5.1
streamlit==1.47.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
tweepy==4.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yfinance==0.2.65
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cmdstanpy==1.2.5
contourpy==1.3.3
curl_cffi==0.12.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.0
frozendict==2.4.6
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
holidays==0.77
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
ml_dtypes==0.5.3
multitasking==0.0.12
namex==0.1.0
narwhals==2.0.1
newsapi-python==0.2.7
nltk==3.9.1
numpy==2.1.3
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
peewee==3.18.2
pillow==11.3.0
platformdirs==4.3.8
plotly==6.2.0
prophet==1.1.7
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.26.0
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
soupsieve==2.7
stanio==0.5.1
streamlit==1.47.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
tweepy==4.16.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
yfinance==0.2.65
